+++ src/web/assets/forms/src/js/formkit/FieldConditions.vue
@@ -263,6 +263,9 @@ export default {
                     testField = field.subfield;
                 }
 
+                // Important to make sure we don't edit the original options, so make un-reactive
+                options = this.clone(options);
+
                 // Special case for agree fields
                 if (testField.type === 'verbb\\formie\\fields\\formfields\\Agree') {
                     return [
@@ -288,7 +291,7 @@ export default {
 
             // Handle submission options which have statically defined options
             if (field && field.valueOptions) {
-                return field.valueOptions;
+                return this.clone(field.valueOptions);
             }
 
             return [];